+++ src/ray/core_worker/common.cc
@@ -17,33 +17,6 @@
 namespace ray {
 namespace core {
 
-std::string WorkerTypeString(WorkerType type) {
-  // TODO(suquark): Use proto3 utils to get the string.
-  if (type == WorkerType::DRIVER) {
-    return "driver";
-  } else if (type == WorkerType::WORKER) {
-    return "worker";
-  } else if (type == WorkerType::SPILL_WORKER) {
-    return "spill_worker";
-  } else if (type == WorkerType::RESTORE_WORKER) {
-    return "restore_worker";
-  }
-  RAY_CHECK(false);
-  return "";
-}
-
-std::string LanguageString(Language language) {
-  if (language == Language::PYTHON) {
-    return "python";
-  } else if (language == Language::JAVA) {
-    return "java";
-  } else if (language == Language::CPP) {
-    return "cpp";
-  }
-  RAY_CHECK(false);
-  return "";
-}
-
 std::string GenerateCachedActorName(const std::string &ns,
                                     const std::string &actor_name) {
   return ns + "-" + actor_name;
